# server/go/api_product.go
/*
 * Product
 *
 * 商品管理デモ。OpenAPI Specification については[github](https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.4.md#schema-1)を参照。
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

import (
	"encoding/json"
	"net/http"
	"strconv"

	"github.com/gorilla/mux"
)

func GetProductById(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")

	vars := mux.Vars(r)
	id := int32(1)
	if s, ok := vars["id"]; ok {
		if n, err := strconv.Atoi(s); err == nil {
			id = int32(n)
		}
	}

	product := Product{}
	switch id {
	case 1:
		product = Product{Id: 1, Name: "商品A", Price: 1000}
	case 2:
		product = Product{Id: 2, Name: "商品B", Price: 2000}
	case 3:
		product = Product{Id: 3, Name: "商品C", Price: 3000}
	default:
		writeCommonError(w, http.StatusNotFound, "A3301", "Product not found")
		return
	}

	w.WriteHeader(http.StatusOK)
	if err := json.NewEncoder(w).Encode(product); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
	}
}

func GetProducts(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")

	products := []Product{
		{Id: 1, Name: "商品A", Price: 1000},
		{Id: 2, Name: "商品B", Price: 2000},
		{Id: 3, Name: "商品C", Price: 3000},
	}

	w.WriteHeader(http.StatusOK)
	if err := json.NewEncoder(w).Encode(products); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
	}
}

func writeCommonError(w http.ResponseWriter, status int, code string, message string) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(status)
	json.NewEncoder(w).Encode(CommonError{
		Code:    code,
		Message: message,
	})
}