# server/go/model_product.go
/*
 * Product
 *
 * 商品管理デモ。OpenAPI Specification については[github](https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.4.md#schema-1)を参照。
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Product struct {

	Id int32 `json:"id"`

	Name string `json:"name"`

	Price int32 `json:"price"`
}